rename file and remove unused method
