+++ eva_cttv_pipeline/clinvar_xml_io/clinvar_xml_io/clinvar_measure.py
@@ -2,7 +2,7 @@ import logging
 from functools import cached_property
 
 from eva_cttv_pipeline.clinvar_xml_io.clinvar_xml_io.hgvs_variant import HgvsVariant
-from eva_cttv_pipeline.clinvar_xml_io.clinvar_xml_io.utils import find_elements, find_optional_unique_element
+from eva_cttv_pipeline.clinvar_xml_io.clinvar_xml_io.xml_parsing import find_elements, find_optional_unique_element
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
